+++ src/content-scripts/keyboard-shortcut/index.js
@@ -1,0 +1,5 @@
+import(
+  browser.runtime.getURL(
+    "/src/content-scripts/keyboard-shortcut/keyboard-shortcut.js",
+  )
+);