+++ CHANGELOG.md
@@ -8,6 +8,8 @@ _released 00 September 2025_
   - Assignments that have titles that start or end with newlines will get those
     trimmed. No more needless empty lines! (looking at you, $GUSH_TEACHER)
   - Add a button to reset all settings.
+  - Press `alt-a` (or `option-a` on Macs) to navigate to the assignment center.
+    (This can be disabled in settings. Defaults to on.)
 - Improvements:
   - Graded assignments will have a green outline around them. The default
     background color for graded assignments is reverted to be the same as